data type from np.double to np.single

+++ mainfunc/converter.py
@@ -8,7 +8,7 @@ def convert(dir_path, sample_rate, n_electrodes, sep):
     dat_path = dir_path + sep + 'temp_wh.dat'   # kilosort outputs waveforms into temp_wh.dat in the working directory
     # channel_map = np.load(dir_path + sep + 'channel_map.npy')
     spike_clusters = np.load(dir_path + sep + 'spike_clusters.npy')
-    spike_times = np.double(np.load(dir_path + sep + 'spike_times.npy'))
+    spike_times = np.single(np.load(dir_path + sep + 'spike_times.npy'))
     n_clusters = len(np.unique(spike_clusters))
 
     sc = SimpleController(dir_path, dat_path, sample_rate, n_electrodes)
@@ -23,7 +23,7 @@ def convert(dir_path, sample_rate, n_electrodes, sep):
     for cluster_id in range(n_clusters):
         try:
             waveforms, spike_ids = sc._get_waveforms(cluster_id)
-            waveformsL.append(waveforms.data[:,:,channel_rank])
+            waveformsL.append(np.single(waveforms.data[:,:,channel_rank]))
             spike_idsL.append(spike_ids)
             primary_electrodeL.append(waveforms.channel_ids[0])
         except:
